+++ iosApp/iosApp/UI/Screen/Home/HomeScreen.swift
@@ -13,7 +13,7 @@ struct HomeScreen: View {
                 TabModel(
                     title: "Home",
                     image: "house",
-                    destination: TabItem {
+                    destination: TabItem { _ in
                         HomeTimelineScreen()
                             .if(horizontalSizeClass != .compact, transform: { view in
                                 view
@@ -51,14 +51,24 @@ struct HomeScreen: View {
                 TabModel(
                     title: "Notification",
                     image: "bell",
-                    destination: TabItem {
+                    destination: TabItem { _ in
                         NotificationScreen()
                     }
                 ),
-                TabModel(title: "Me", image: "person.circle", destination: TabItem {ProfileScreen(userKey: nil)})
+                TabModel(
+                    title: "Me",
+                    image: "person.circle",
+                    destination: TabItem { router in
+                        ProfileScreen(
+                            userKey: nil,
+                            toProfileMedia: { userKey in
+                                router.navigate(to: .profileMedia(userKey: userKey.description()))
+                            }
+                        )
+                    }
+                )
             ],
             secondaryItems: [
-
             ],
             leading: HStack {
                 if case .success(let data) = onEnum(of: viewModel.model.user) {
@@ -117,12 +127,11 @@ class HomeViewModel: MoleculeViewModelBase<ActiveAccountState, ActiveAccountPres
 
 struct TabItem<Content: View>: View {
     @State var router = Router<TabDestination>()
-    let content: () -> Content
-
+    let content: (Router<TabDestination>) -> Content
     var body: some View {
         NavigationStack(path: $router.navPath) {
-            content()
-                .withTabRouter()
+            content(router)
+                .withTabRouter(router: router)
         }.environment(\.openURL, OpenURLAction { url in
             if let event = AppDeepLink.shared.parse(url: url.absoluteString) {
                 switch onEnum(of: event) {
@@ -138,7 +147,6 @@ struct TabItem<Content: View>: View {
                 return .discarded
             }
         })
-
     }
 }
 
@@ -146,11 +154,9 @@ struct TabItem<Content: View>: View {
 class StatusEvent: MastodonStatusEvent, MisskeyStatusEvent, BlueskyStatusEvent {
     let accountRepository = KoinHelper.shared.accountRepository
     var composeStatus: ComposeStatus?
-
     func onReplyClick(status: UiStatus.Mastodon) {
         composeStatus = ComposeStatusReply(statusKey: status.statusKey)
     }
-
     func onReblogClick(status: UiStatus.Mastodon) {
         Task {
             if let account = accountRepository.get(accountKey: status.accountKey) as? UiAccountMastodon {
@@ -158,7 +164,6 @@ class StatusEvent: MastodonStatusEvent, MisskeyStatusEvent, BlueskyStatusEvent {
             }
         }
     }
-
     func onLikeClick(status: UiStatus.Mastodon) {
         Task {
             if let account = accountRepository.get(accountKey: status.accountKey) as? UiAccountMastodon {
@@ -166,7 +171,6 @@ class StatusEvent: MastodonStatusEvent, MisskeyStatusEvent, BlueskyStatusEvent {
             }
         }
     }
-
     func onBookmarkClick(status: UiStatus.Mastodon) {
         Task {
             if let account = accountRepository.get(accountKey: status.accountKey) as? UiAccountMastodon {
@@ -174,10 +178,8 @@ class StatusEvent: MastodonStatusEvent, MisskeyStatusEvent, BlueskyStatusEvent {
             }
         }
     }
-
     func onMediaClick(media: UiMedia) {
     }
-
     func onReportClick(status: UiStatus.Mastodon) {
         Task {
             if let account = accountRepository.get(accountKey: status.accountKey) as? UiAccountMastodon {
@@ -186,7 +188,6 @@ class StatusEvent: MastodonStatusEvent, MisskeyStatusEvent, BlueskyStatusEvent {
             }
         }
     }
-
     func onReactionClick(data: UiStatus.Misskey, reaction: UiStatus.MisskeyEmojiReaction) {
         Task {
             if let account = accountRepository.get(accountKey: data.accountKey) as? UiAccountMisskey {
@@ -194,11 +195,9 @@ class StatusEvent: MastodonStatusEvent, MisskeyStatusEvent, BlueskyStatusEvent {
             }
         }
     }
-
     func onReplyClick(data: UiStatus.Misskey) {
         composeStatus = ComposeStatusReply(statusKey: data.statusKey)
     }
-
     func onReblogClick(data: UiStatus.Misskey) {
         Task {
             if let account = accountRepository.get(accountKey: data.accountKey) as? UiAccountMisskey {
@@ -206,14 +205,11 @@ class StatusEvent: MastodonStatusEvent, MisskeyStatusEvent, BlueskyStatusEvent {
             }
         }
     }
-
     func onQuoteClick(data: UiStatus.Misskey) {
         composeStatus = ComposeStatusQuote(statusKey: data.statusKey)
     }
-
     func onAddReactionClick(data: UiStatus.Misskey) {
     }
-
     func onReportClick(data: UiStatus.Misskey) {
         Task {
             if let account = accountRepository.get(accountKey: data.accountKey) as? UiAccountMisskey {
@@ -221,11 +217,9 @@ class StatusEvent: MastodonStatusEvent, MisskeyStatusEvent, BlueskyStatusEvent {
             }
         }
     }
-
     func onReplyClick(data: UiStatus.Bluesky) {
         composeStatus = ComposeStatusReply(statusKey: data.statusKey)
     }
-
     func onReblogClick(data: UiStatus.Bluesky) {
         Task {
             if let account = accountRepository.get(accountKey: data.accountKey) as? UiAccountBluesky {
@@ -233,11 +227,9 @@ class StatusEvent: MastodonStatusEvent, MisskeyStatusEvent, BlueskyStatusEvent {
             }
         }
     }
-
     func onQuoteClick(data: UiStatus.Bluesky) {
         composeStatus = ComposeStatusQuote(statusKey: data.statusKey)
     }
-
     func onLikeClick(data: UiStatus.Bluesky) {
         Task {
             if let account = accountRepository.get(accountKey: data.accountKey) as? UiAccountBluesky {
@@ -245,7 +237,6 @@ class StatusEvent: MastodonStatusEvent, MisskeyStatusEvent, BlueskyStatusEvent {
             }
         }
     }
-
     func onReportClick(data: UiStatus.Bluesky, reason: BlueskyReportStatusStateReportReason) {
         Task {
             if let account = accountRepository.get(accountKey: data.accountKey) as? UiAccountBluesky {
@@ -253,7 +244,6 @@ class StatusEvent: MastodonStatusEvent, MisskeyStatusEvent, BlueskyStatusEvent {
             }
         }
     }
-
     func onDeleteClick(accountKey: MicroBlogKey, statusKey: MicroBlogKey) {
         Task {
             if let account = accountRepository.get(accountKey: accountKey) {
