+++ iosApp/iosApp/UI/Screen/Profile/Mastodon.swift
@@ -1,0 +1,92 @@
+import SwiftUI
+import shared
+
+struct MastodonFollowButton: View {
+    let relation: UiState<UiRelation>
+    let isMe: UiState<KotlinBoolean>
+    let onFollowClick: (UiRelation) -> Void
+    var body: some View {
+        if case .success(let success) = onEnum(of: isMe),
+            !success.data.boolValue,
+           case .success(let relationData) = onEnum(of: relation),
+           let mastodonRelation = relationData.data as? UiRelationMastodon {
+            let text = if mastodonRelation.blocking {
+                "Blocked"
+            } else if mastodonRelation.following {
+                "Following"
+            } else if mastodonRelation.requested {
+                "Requested"
+            } else {
+                "Follow"
+            }
+            Button(action: {
+                onFollowClick(mastodonRelation)
+            }, label: {
+                Text(text)
+            })
+            .buttonStyle(.borderless)
+        } else {
+            EmptyView()
+        }
+    }
+}
+
+struct MastodonMenu: View {
+    let relation: UiRelationMastodon
+    let onMuteClick: () -> Void
+    let onBlockClick: () -> Void
+    var body: some View {
+        Button(action: onMuteClick, label: {
+            let text = if relation.muting {
+                "Unmute"
+            } else {
+                "Mute"
+            }
+            let icon = if relation.muting {
+                "speaker"
+            } else {
+                "speaker.slash"
+            }
+            Label(text, systemImage: icon)
+        })
+        Button(action: onBlockClick, label: {
+            let text = if relation.blocking {
+                "Unblock"
+            } else {
+                "Block"
+            }
+            let icon = if relation.blocking {
+                "xmark.circle"
+            } else {
+                "checkmark.circle"
+            }
+            Label(text, systemImage: icon)
+        })
+    }
+}
+
+struct MastodonProfileHeader: View {
+    let user: UiUser.Mastodon
+    let relation: UiState<UiRelation>
+    let isMe: UiState<KotlinBoolean>
+    let onFollowClick: (UiRelation) -> Void
+    var body: some View {
+        CommonProfileHeader(
+            bannerUrl: user.bannerUrl,
+            avatarUrl: user.avatarUrl,
+            displayName: user.extra.nameMarkdown,
+            handle: user.handle,
+            description: user.extra.descriptionMarkdown,
+            headerTrailing: {
+                MastodonFollowButton(relation: relation, isMe: isMe, onFollowClick: onFollowClick)
+            }, content: {
+                VStack(alignment: .leading) {
+                    MatrixView(
+                        followCount: user.matrices.followsCountHumanized,
+                        fansCount: user.matrices.fansCountHumanized
+                    )
+                    FieldsView(fields: user.extra.fieldsMarkdown)
+                }
+            })
+    }
+}
